refactor: NodeList
