--- src/galaxy/nodeDetails/NodeList.jsx
@@ -1,57 +1,0 @@
-/**
- * This component renders list of packages in single window.
- * There can be multiple lists opened at any time. The list is currently
- * managed by windowCollectionView
- */
-import React from "react";
-import ReactList from "react-list";
-import WindowTitle from "./WindowTitle.jsx";
-import { NodeListItem } from "./NodeInfoRow.jsx";
-
-var windowId = 0;
-
-export const NodeList = ({ windowViewModel }) => {
-  windowId += 1;
-
-  var className = "window-container";
-
-  if (windowViewModel.className) {
-    className += " " + windowViewModel.className;
-  }
-
-  var items = windowViewModel.list;
-  var id = windowId + windowViewModel.className + items.length;
-
-  return (
-    <div className={className}>
-      <WindowTitle viewModel={windowViewModel} />
-      <div className="window-list-content">{content(items)}</div>
-    </div>
-  );
-
-  function renderItem(idx, key) {
-    var vm = items[idx];
-    return <NodeListItem key={key} viewModel={vm} />;
-  }
-
-  function getHeight() {
-    // FIXME: Hardcoding is not good.
-    return 20;
-  }
-
-  function content(items) {
-    if (items.length > 0) {
-      return (
-        <ReactList
-          itemRenderer={renderItem}
-          length={items.length}
-          itemSizeGetter={getHeight}
-          type="variable"
-          key={id}
-        />
-      );
-    } else {
-      return null;
-    }
-  }
-};
